rest/requirements.txt
Flask
gunicorn
python-dotenv
redis
flask_hal
pyjwt
uuid